work on basic queue service
