+++ src/startup/rabbitmq.js
@@ -9,7 +9,6 @@ async () => {
     try {
         connectRabbitMQ = await amqp.connect('amqp://localhost');
         const channel = await connectRabbitMQ.createChannel();
-        await channel.assertQueue('todoTasks');
         container.cradle.logger.info("Rabbitmq connected Successfully")
         return { connectRabbitMQ, channel };
     } catch (error) {
